refactor: 같은 목적의 class 를 한 package 로 합치고, 명시적인 이름으로 변경
