+++ src/main/java/chess/db/DBConnector.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.db;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
