+++ src/test/java/chess/db/ChessGameDaoTest.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.db;
 
 import static org.assertj.core.api.Assertions.*;
